account for NBS detune (pitch), clamp final sound value between 0.5 and 2 as per wiki.vg

+++ commons/src/main/java/org/phantazm/commons/MathUtils.java
@@ -36,10 +36,9 @@ public final class MathUtils {
         return Math.min(Math.max(value, min), max);
     }
 
-    public static int clamp(int value, int min, int max) {
+    public static float clamp(float value, float min, float max) {
         return Math.min(Math.max(value, min), max);
     }
-
 
     /**
      * Computes the lambertW (product logarithm) for a given real value and branch.
